+++ cynic-codegen/src/fragment_derive/input.rs
@@ -89,10 +89,13 @@ pub struct FragmentDeriveField {
     pub(super) attrs: Vec<syn::Attribute>,
 
     #[darling(default)]
-    pub(super) flatten: bool,
+    pub(super) flatten: SpannedValue<bool>,
 
     #[darling(default)]
     pub(super) recurse: Option<SpannedValue<u8>>,
+
+    #[darling(default)]
+    pub(super) spread: SpannedValue<bool>,
 
     #[darling(default)]
     rename: Option<SpannedValue<String>>,
@@ -100,10 +103,26 @@ pub struct FragmentDeriveField {
 
 impl FragmentDeriveField {
     pub fn validate(&self) -> Result<(), Errors> {
-        if self.flatten && self.recurse.is_some() {
+        if *self.flatten && self.recurse.is_some() {
             return Err(syn::Error::new(
                 self.recurse.as_ref().unwrap().span(),
                 "A field can't be recurse if it's being flattened",
+            )
+            .into());
+        }
+
+        if *self.flatten && *self.spread {
+            return Err(syn::Error::new(
+                self.flatten.span(),
+                "A field can't be flattened if it's also being spread",
+            )
+            .into());
+        }
+
+        if *self.spread && self.recurse.is_some() {
+            return Err(syn::Error::new(
+                self.recurse.as_ref().unwrap().span(),
+                "A field can't be recurse if it's being spread",
             )
             .into());
         }
@@ -112,10 +131,12 @@ impl FragmentDeriveField {
     }
 
     pub fn type_check_mode(&self) -> CheckMode {
-        if self.flatten {
+        if *self.flatten {
             CheckMode::Flattening
         } else if self.recurse.is_some() {
             CheckMode::Recursing
+        } else if *self.spread {
+            CheckMode::Spreading
         } else {
             CheckMode::Normal
         }
@@ -148,24 +169,36 @@ mod tests {
                         ident: Some(format_ident!("field_one")),
                         ty: syn::parse_quote! { String },
                         attrs: vec![],
-                        flatten: false,
+                        flatten: false.into(),
                         recurse: None,
+                        spread: false.into(),
                         rename: None,
                     },
                     FragmentDeriveField {
                         ident: Some(format_ident!("field_two")),
                         ty: syn::parse_quote! { String },
                         attrs: vec![],
-                        flatten: true,
+                        flatten: true.into(),
                         recurse: None,
+                        spread: false.into(),
                         rename: None,
                     },
                     FragmentDeriveField {
                         ident: Some(format_ident!("field_three")),
                         ty: syn::parse_quote! { String },
                         attrs: vec![],
-                        flatten: false,
+                        flatten: false.into(),
                         recurse: Some(8.into()),
+                        spread: false.into(),
+                        rename: Some("fieldThree".to_string().into()),
+                    },
+                    FragmentDeriveField {
+                        ident: Some(format_ident!("some_spread")),
+                        ty: syn::parse_quote! { String },
+                        attrs: vec![],
+                        flatten: false.into(),
+                        recurse: None,
+                        spread: true.into(),
                         rename: Some("fieldThree".to_string().into()),
                     },
                 ],
@@ -191,24 +224,45 @@ mod tests {
                         ident: Some(format_ident!("field_one")),
                         ty: syn::parse_quote! { String },
                         attrs: vec![],
-                        flatten: false,
+                        flatten: false.into(),
                         recurse: None,
+                        spread: false.into(),
                         rename: None,
                     },
                     FragmentDeriveField {
                         ident: Some(format_ident!("field_two")),
                         ty: syn::parse_quote! { String },
                         attrs: vec![],
-                        flatten: true,
+                        flatten: true.into(),
                         recurse: Some(8.into()),
+                        spread: false.into(),
                         rename: None,
                     },
                     FragmentDeriveField {
                         ident: Some(format_ident!("field_three")),
                         ty: syn::parse_quote! { String },
                         attrs: vec![],
-                        flatten: true,
+                        flatten: true.into(),
                         recurse: Some(8.into()),
+                        spread: false.into(),
+                        rename: None,
+                    },
+                    FragmentDeriveField {
+                        ident: Some(format_ident!("some_spread")),
+                        ty: syn::parse_quote! { String },
+                        attrs: vec![],
+                        flatten: true.into(),
+                        recurse: None,
+                        spread: true.into(),
+                        rename: None,
+                    },
+                    FragmentDeriveField {
+                        ident: Some(format_ident!("some_other_spread")),
+                        ty: syn::parse_quote! { String },
+                        attrs: vec![],
+                        flatten: false.into(),
+                        recurse: Some(8.into()),
+                        spread: true.into(),
                         rename: None,
                     },
                 ],
@@ -221,7 +275,7 @@ mod tests {
         };
 
         let errors = input.validate().unwrap_err();
-        assert_eq!(errors.len(), 2);
+        assert_eq!(errors.len(), 4);
     }
 
     #[test]
@@ -257,24 +311,27 @@ mod tests {
                         ident: Some(format_ident!("field_one")),
                         ty: syn::parse_quote! { String },
                         attrs: vec![],
-                        flatten: false,
+                        flatten: false.into(),
                         recurse: None,
+                        spread: false.into(),
                         rename: None,
                     },
                     FragmentDeriveField {
                         ident: Some(format_ident!("field_two")),
                         ty: syn::parse_quote! { String },
                         attrs: vec![],
-                        flatten: true,
+                        flatten: true.into(),
                         recurse: None,
+                        spread: false.into(),
                         rename: None,
                     },
                     FragmentDeriveField {
                         ident: Some(format_ident!("field_three")),
                         ty: syn::parse_quote! { String },
                         attrs: vec![],
-                        flatten: false,
+                        flatten: false.into(),
                         recurse: Some(8.into()),
+                        spread: false.into(),
                         rename: None,
                     },
                 ],