+++ examples/examples/starwars.rs
@@ -59,7 +59,7 @@ mod test {
     use super::*;
 
     #[test]
-    fn snapshot_test_menu_query() {
+    fn snapshot_test_query() {
         // Running a snapshot test of the query building functionality as that gives us
         // a place to copy and paste the actual GQL we're using for running elsewhere,
         // and also helps ensure we don't change queries by mistake